requirements.txt
cycler==0.10.0
fbs-tutorial==1.0.0
fbs-tutorial-shim==1.0.2
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.2
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
six==1.16.0
